pipeline/requirements.txt
pandas
spacy
pycountry
rapidfuzz
boto3
python-dotenv
pandarallel